feat: めんどうくせえ

+++ EnglishApp.Domain/Entities/UserProfileInsertEntity.cs
@@ -2,7 +2,7 @@
 
 public sealed class UserProfileInsertEntity(int userId,
                                            string nickName,
-                                           byte gender,
+                                           int gender,
                                            int gradeId,
                                            int learningPurposeId,
                                            int prefectureId,
@@ -12,7 +12,7 @@ public sealed class UserProfileInsertEntity(int userId,
 {
     public int UserId { get; } = userId;
     public string NickName { get; } = nickName;
-    public byte Gender { get; } = gender;
+    public int Gender { get; } = gender;
     public int GradeId { get; } = gradeId;
     public int LearningPurposeId { get; } = learningPurposeId;
     public int PrefectureId { get; } = prefectureId;
